% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_account}
\alias{chime_get_account}
\title{Retrieves details for the specified Amazon Chime account, such as
account type and supported licenses}
\usage{
chime_get_account(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Retrieves details for the specified Amazon Chime account, such as
account type and supported licenses.
}
\section{Request syntax}{
\preformatted{chime$get_account(
  AccountId = "string"
)
}
}

\keyword{internal}
