% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_invite_users}
\alias{chime_invite_users}
\title{Sends email invites to as many as 50 users, inviting them to the
specified Amazon Chime Team account}
\usage{
chime_invite_users(AccountId, UserEmailList)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserEmailList}{[required] The user email addresses to which to send the invite.}
}
\description{
Sends email invites to as many as 50 users, inviting them to the
specified Amazon Chime \code{Team} account. Only \code{Team} account types are
currently supported for this action.
}
\section{Request syntax}{
\preformatted{chime$invite_users(
  AccountId = "string",
  UserEmailList = list(
    "string"
  )
)
}
}

\keyword{internal}
