% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_global_settings}
\alias{chime_update_global_settings}
\title{Updates global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings}
\usage{
chime_update_global_settings(BusinessCalling, VoiceConnector)
}
\arguments{
\item{BusinessCalling}{[required] The Amazon Chime Business Calling settings.}

\item{VoiceConnector}{[required] The Amazon Chime Voice Connector settings.}
}
\description{
Updates global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings.
}
\section{Request syntax}{
\preformatted{chime$update_global_settings(
  BusinessCalling = list(
    CdrBucket = "string"
  ),
  VoiceConnector = list(
    CdrBucket = "string"
  )
)
}
}

\keyword{internal}
